+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Intermediate work on Appointments UI [#2464](https://github.com/rokwire/illinois-app/issues/2464).
 ### Changed
 - Demographics Questionnaire renamed to Research, various UX updates applied [#2465](https://github.com/rokwire/illinois-app/issues/2465).